+++ packages/ui/src/components/Table/components/Column.tsx
@@ -25,11 +25,11 @@ import { RiArrowUpLine, RiArrowDownLine } from '@remixicon/react';
 /** @public */
 export const Column = (props: ColumnProps) => {
   const { classNames, cleanedProps } = useStyles(TableDefinition, props);
-  const { children, ...rest } = cleanedProps;
+  const { className, children, ...rest } = cleanedProps;
 
   return (
     <ReactAriaColumn
-      className={clsx(classNames.head, styles[classNames.head])}
+      className={clsx(classNames.head, styles[classNames.head], className)}
       {...rest}
     >
       {({ allowsSorting, sortDirection }) => (
@@ -48,10 +48,10 @@ export const Column = (props: ColumnProps) => {
                 styles[classNames.headSortButton],
               )}
             >
-              {sortDirection === 'ascending' ? (
-                <RiArrowUpLine size={16} />
-              ) : (
+              {sortDirection === 'descending' ? (
                 <RiArrowDownLine size={16} />
+              ) : (
+                <RiArrowUpLine size={16} />
               )}
             </span>
           )}