+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @backstage/cli
 
+## 0.34.6-next.1
+
+### Patch Changes
+
+- 7fbac5c: Updated to use new utilities from `@backstage/cli-common`.
+- Updated dependencies
+  - @backstage/cli-node@0.2.16-next.1
+  - @backstage/cli-common@0.1.16-next.1
+
+## 0.34.6-next.0
+
+### Patch Changes
+
+- c8c2329: Add proxy configuration from env-vars to create-app tasks
+- Updated dependencies
+  - @backstage/integration@1.18.3-next.0
+  - @backstage/cli-common@0.1.16-next.0
+  - @backstage/cli-node@0.2.16-next.0
+  - @backstage/config@1.3.6
+  - @backstage/release-manifests@0.0.13
+  - @backstage/config-loader@1.10.7-next.0
+  - @backstage/catalog-model@1.7.6
+  - @backstage/errors@1.2.7
+  - @backstage/eslint-plugin@0.2.0
+  - @backstage/types@1.2.2
+
 ## 0.34.5
 
 ### Patch Changes