+++ packages/codemods/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/codemods
 
+## 0.1.53-next.1
+
+### Patch Changes
+
+- 688f070: Updated to use new utilities from `@backstage/cli-common`.
+- Updated dependencies
+  - @backstage/cli-common@0.1.16-next.1
+
+## 0.1.53-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/cli-common@0.1.16-next.0
+
 ## 0.1.52
 
 ### Patch Changes