+++ plugins/scaffolder-backend-module-gitea/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-scaffolder-backend-module-gitea
 
+## 0.2.16-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@1.5.1-next.0
+  - @backstage/integration@1.18.3-next.0
+  - @backstage/plugin-scaffolder-node@0.12.2-next.0
+  - @backstage/config@1.3.6
+  - @backstage/errors@1.2.7
+
 ## 0.2.15
 
 ### Patch Changes